backend/requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.10.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
flake8==4.0.1
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
Werkzeug==2.0.3
